refactor : removed redundancy in CMakeList and moved rapidobj.hpp to parser directory
