+++ CMakeLists.txt
@@ -17,11 +17,8 @@ file(GLOB_RECURSE HEADERS "src/*.h")
 file(GLOB_RECURSE SRCS "src/*.cpp")
 
 # Add static library for pySunlight
-# add_library(sunlight ${HEADERS} ${SRCS})
-add_executable(sunlight ${SRCS} ${HEADERS}
-        src/parsers/ObjParser.cpp
-        src/parsers/ObjParser.h
-        src/tests/intersection_test.h)
+#add_library(sunlight ${HEADERS} ${SRCS})
+ add_executable(sunlight ${SRCS} ${HEADERS})
 
 target_include_directories(sunlight PRIVATE "src/")
